Add NonNullable return type requirement for `Maybe.map`.
This updates the typings file to be consistent with the javascript which
prevents `null` values from being initialized within a `Maybe`.

The `NonNullable` type requires TypeScript 2.8+.

Additionally, the TS update brings better type coverage which caused one
`NEL` test to fail and has been updated to reflect the correct type.